+++ bam-ui/src/App.js
@@ -8,9 +8,15 @@ import NotFound from "./pages/notFound/NotFound";
 import BackOffice from "./pages/dashboard/BackOffice";
 import Layout from "./forms/formlayout/LoginRegister";
 
+import Footer from "./features/footer/Footer";
+import Header from "./features/header/Header";
+import Navbar from "./features/navbar/Navbar";
+
 function App() {
     return (
         <div className="App">
+            <Navbar />
+            <Header page="auctions" />
             <Routes>
                 <Route path="/" element={<Home />} />
                 <Route path="/login/register" element={<Layout />} />
@@ -19,6 +25,7 @@ function App() {
                 <Route path="/dashboard" element={<BackOffice />} />
                 <Route path="*" element={<NotFound />} />
             </Routes>
+            <Footer />
         </div>
     );
 }
