--- bam-ui/src/features/header/headerSearch.module.css
@@ -1,54 +1,0 @@
-.header_search {
-    
-    position: relative;
-}
-.header_search_container {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    border: 4px solid var(--g);
-    background-color: white;
-    height:45px;
-    width: 40%;
-    max-width: 1024px;
-    position: absolute;
-    bottom: -25px;
-}
-.header_search_section{
-    display: flex;
-    width: 90px;
-    align-items: center;
-    gap:0.8rem;
-    padding-left: 0.5rem;
-    cursor: pointer;
-    order:2;
-}
-.header_search_btn {
-   border: var(--g);
-   color:var(--d);
-   background-color: var(--g);
-   padding: 8px 5px;
-   width: 100%;
-   order: 1;
-   ;
-}
-.header_search_btn:hover {
-    color:var(--gy);
-    background: var(--d);
-}
-
-.header_search_icon {
-    color:lightgray;
-    font-size: 1.3rem;
-}
-.header_search_input {
-    border: none;
-    cursor: pointer;
-    outline: none;
-    font-size: 1.1rem;
-}
-
-
-
-
-
