+++ bam-ui/src/pages/home/Home.js
@@ -1,5 +1,5 @@
 //import FeaturedCity from "../../features/city/FeaturedCity";
-import Footer from "../../features/footer/Footer";
+//import Footer from "../../features/footer/Footer";
 import Header from "../../features/header/Header";
 import Navbar from "../../features/navbar/Navbar";
 //import PropertyType from "../../features/property/PropertyType";
@@ -7,7 +7,7 @@ import styles from "./Home.module.css";
 import { useState } from "react";
 import { useSearchParams, useNavigate, useParams } from "react-router-dom";
 
-const Home = () => {
+const Home = ({ page }) => {
     const navigate = useNavigate();
     let city = useParams();
     const [citySearch, setCitySearch] = useState("");
@@ -18,7 +18,6 @@ const Home = () => {
 
     return (
         <>
-            <Navbar />
             <Header />
             {/*------------------------------------ Main cities top 3---------------------------------------------------------------- */}
             <div className={styles.home_main_container}>
@@ -179,7 +178,6 @@ const Home = () => {
                 </div>
             </div>
             {/*------------------------- Footer -------------------------------------------------------- */}
-            <Footer />
         </>
     );
 };
